pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aita"
version = "0.1.0"
description = "AitA - AI Travel Assistant"
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv",
  "requests",
  # ... other dependencies
]

[project.scripts]
aita = "aita.main:main"

[tool.hatch.envs.default]
dependencies = ["pytest", "ipython"]

[tool.hatch.envs.default.scripts]
test = "pytest test"
